pseiferth-libxtreemfsjava: New filesize of files is now merged into stat object from metadatacache. Added missing exceptions. Fixed failed test and created new tests for client and uuidresolver. Test coverage: 60%.
git-svn-id: https://xtreemfs.googlecode.com/svn/branches/pseiferth-libxtreemfsjava@2806 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ java/servers/src/org/xtreemfs/common/libxtreemfs/Options.java
@@ -77,7 +77,7 @@ public class Options {
     /**
      * Maximum number of entries of the StatCache. Default: 100000
      */
-    private long    metadataCacheSize     = 100000;
+    private int    metadataCacheSize     = 100000;
 
     /**
      * Time to live for MetadataCache entries. Default: 120
@@ -138,7 +138,7 @@ public class Options {
      */
     private int     lingerTimeout_s       = 600;
 
-    public long getMetadataCacheSize() {
+    public int getMetadataCacheSize() {
         return metadataCacheSize;
     }
